+++ libs/wal_decoder/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod decoder;
+pub mod models;
